hk_back/hk_back/users/views.py
from django.core.serializers import serialize
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework import status
from rest_framework.exceptions import AuthenticationFailed

from .models import User
from .serializers import UserSerializer

import jwt, datetime

# Create your views here.
class RegisterView(APIView):
    def post(self, request):
        serializer = UserSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        serializer.save()

        response = Response()
        response.data = {
            'status': status.HTTP_200_OK,
            'message': 'User registered successfully',
            'data': serializer.data
        }
        return response